Refactor blur handling into reusable component
